requirements.txt
twilio>=8.10.0
flask>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.3.0

